+++ src/NonEmptyString.ts
@@ -25,10 +25,10 @@ import {
 } from "fp-ts/string"
 import * as Str from "./String"
 import { pack, unpack, over } from "./Newtype"
-import { not } from "fp-ts/Predicate"
+import { Predicate, not } from "fp-ts/Predicate"
 import { Endomorphism } from "fp-ts/Endomorphism"
 import type { Show as TShow } from "fp-ts/Show"
-import { ReadonlyNonEmptyArray } from "fp-ts/lib/ReadonlyNonEmptyArray"
+import { ReadonlyNonEmptyArray } from "fp-ts/ReadonlyNonEmptyArray"
 
 type Show<A> = TShow<A>
 
@@ -216,6 +216,7 @@ export const reverse: Endomorphism<NonEmptyString> = over(Str.reverse)
 /**
  * Calculate the number of characters in a `NonEmptyString`.
  *
+ * @category 3 Functions
  * @since 0.17.0
  */
 export const size: (x: NonEmptyString) => number = flow(toString, _size)
@@ -223,20 +224,20 @@ export const size: (x: NonEmptyString) => number = flow(toString, _size)
 /**
  * Split a `NonEmptyString` into an array of strings using `separator`.
  *
+ * @category 3 Functions
  * @since 0.17.0
  */
-export const split: (
+export const split = (
   separator: string | RegExp,
-) => (s: NonEmptyString) => ReadonlyNonEmptyArray<string> = separator =>
+): ((x: NonEmptyString) => ReadonlyNonEmptyArray<string>) =>
   flow(toString, _split(separator))
 
 /**
- * Returns a `boolean` if a `NonEmptyString` contains `needle`.
+ * Predicate upon the presence of a `search` string anywhere in a
+ * `NonEmptyString`.
  *
+ * @category 3 Functions
  * @since 0.17.0
  */
-export const includes: (
-  needle: string,
-  pos?: number,
-) => (haystack: NonEmptyString) => boolean = (needle, pos) =>
-  flow(toString, _includes(needle, pos))
+export const includes = (search: string): Predicate<NonEmptyString> =>
+  flow(toString, _includes(search))